Move Header back to pages
