+++ src/pages/pricing/Pricing.module.css
