+++ src/pages/profile/Profile.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ProfileInfo from "./ProfileInfo/ProfileInfo";
+import ProfileInfo from "../../components/Profile/ProfileInfo/ProfileInfo";
 
 const Profile = () => {
   return (
