+++ src/routes/ProtectedRoute.jsx
@@ -1,12 +1,11 @@
 // components/ProtectedRoute.jsx
 import React from "react";
 import { Outlet } from "react-router-dom";
-import { useAuth } from "../components/Authentication/hooks/use-auth";
+import { useAuth } from "../components/Authentication/hooks/useAuth";
 import AuthPopup from "../components/Authentication/components/AuthPopup";
 
 const ProtectedRoute = () => {
   const auth = useAuth();
-  console.log("ProtectedRoute - Auth state:", auth);
   return auth.isAuth ? <Outlet /> : <AuthPopup />;
 };
